Separate params package into licenses and versioning
