+++ jsonrpc/web3_endpoint.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/0xPolygon/polygon-edge/helper/hex"
 	"github.com/0xPolygon/polygon-edge/helper/keccak"
-	"github.com/0xPolygon/polygon-edge/params"
+	"github.com/0xPolygon/polygon-edge/versioning"
 )
 
 // Web3 is the web3 jsonrpc endpoint
@@ -13,7 +13,7 @@ type Web3 struct{}
 
 // ClientVersion returns the version of the web3 client (web3_clientVersion)
 func (w *Web3) ClientVersion() (interface{}, error) {
-	return fmt.Sprintf("polygon-edge [%s]", params.Version), nil
+	return fmt.Sprintf("polygon-edge [%s]", versioning.Version), nil
 }
 
 // Sha3 returns Keccak-256 (not the standardized SHA3-256) of the given data
