+++ versioning/versioning.go
@@ -1,4 +1,4 @@
-package params
+package versioning
 
 var (
 	// Version is the main version at the moment.
